message send errors: send success/failure events
This commit adds the ability to send custom success/failure events when
messages succeed or fail to bridge. The event type is
"com.beeper.message_send_status".

To enable, set bridge.send_message_send_status_events to true.

This commit also adds the ability to control whether the error notices
are sent in addition to the custom events.